+++ validation/core_validator.py
@@ -73,6 +73,8 @@ class CoreValidator:
 
         _my_stake = self.metagraph.S[self.metagraph.hotkeys.index(self.public_hotkey_address)]
         self._my_prop_of_stake = (_my_stake / sum(self.metagraph.S)).item()
+        if is_testnet:
+            self._my_prop_of_stake = 1.0
 
         validation_utils.connect_to_external_server()
 
@@ -118,13 +120,14 @@ class CoreValidator:
             Task.dreamshaper_image_to_image: 0.05,
             Task.jugger_inpainting: 0.05,
             Task.clip_image_embeddings: 0.0,
-            Task.avatar: 0.2,
+            Task.avatar: 0.20,
         }
         db_manager.task_weights = weights
         return weights
 
-    def _correct_capacities(self) -> None:
+    async def _post_and_correct_capacities(self) -> None:
         self._correct_for_max_capacities()
+        await self._post_miner_capacities_to_tauvision()
         self._correct_capacities_for_my_stake()
 
     def _correct_capacities_for_my_stake(self) -> None:
@@ -192,9 +195,7 @@ class CoreValidator:
                         continue
                     if uid not in self.capacities_for_tasks[task]:
                         self.capacities_for_tasks[task][uid] = float(volume.volume)
-            self._correct_capacities()
-        capacities_to_log = {k.value: v for k, v in self.capacities_for_tasks.items()}
-        bt.logging.info(f"Capacities: {capacities_to_log}")
+            await self._post_and_correct_capacities()
         bt.logging.info("Done fetching available tasks!")
 
     async def resync_metagraph(self):
@@ -246,6 +247,31 @@ class CoreValidator:
             data_type_to_post=post_stats.DataTypeToPost.MINER_CAPACITIES,
         )
 
+
+    def _post_uid_records_to_tauvision(self) -> None:
+        data_to_post = []
+        for task in self.uid_manager.uid_records_for_tasks:
+            for record in self.uid_manager.uid_records_for_tasks[task].values():
+                data_to_post.append(post_stats.UidRecordPostObject(
+                    miner_hotkey=record.hotkey,
+                    validator_hotkey=self.public_hotkey_address,
+                    axon_uid=record.axon_uid,
+                    period_score=record.period_score,
+                    declared_volume=record.declared_volume,
+                    consumed_volume=record.consumed_volume,
+                    requests_429=record.requests_429,
+                    requests_500=record.requests_500,
+                    total_requests_made=record.total_requests_made,
+                    task=task,
+                ))
+
+        post_data = post_stats.UidRecordsPostBody(data=data_to_post)
+        asyncio.create_task(post_stats.post_to_tauvision(
+            data_to_post=post_data.dump(),
+            keypair=self.keypair,
+            data_type_to_post=post_stats.DataTypeToPost.UID_RECORD,
+        ))
+
     async def run_vali(self) -> None:
         iteration = 1
         while True:
@@ -258,12 +284,11 @@ class CoreValidator:
                 data_type_to_post=post_stats.DataTypeToPost.VALIDATOR_INFO,
             )
 
-            db_manager.delete_data_older_than_date(minutes=60 * 24)
-            db_manager.delete_tasks_older_than_date(minutes=120)
+            await db_manager.delete_data_older_than_date(minutes=60 * 24 * 2)
+            await db_manager.delete_tasks_older_than_date(minutes=120)
 
             # Wait for initial syncing of metagraph
             await self.resync_metagraph()
-            await self._post_miner_capacities_to_tauvision()
             self.scorer.start_scoring_results_if_not_already()
 
             bt.logging.info("🚀 Starting to score stuff, metagraph is synced...")
@@ -277,7 +302,8 @@ class CoreValidator:
             await self.uid_manager.start_synthetic_scoring()
             await self.uid_manager.collect_synthetic_scoring_results()
             self.uid_manager.calculate_period_scores_for_uids()
-            self.uid_manager.store_period_scores()
+            self._post_uid_records_to_tauvision()
+            await  self.uid_manager.store_period_scores()
 
             bt.logging.info(f"Finished scoring for iteration: {iteration}. Now settings weights")
             iteration += 1