+++ validation/proxy/api_server/asgi.py
@@ -1,24 +1,24 @@
-from fastapi import FastAPI
-from fastapi import Request
+from fastapi import FastAPI, Request, Response
+from fastapi.responses import JSONResponse
+from starlette import status
 import uvicorn
 import asyncio
+import aiosqlite
 from config.validator_config import config as validator_config
 from validation.proxy.api_server.image.endpoints import router as image_router
 from validation.proxy.api_server.text.endpoints import router as text_router
 from validation.core_validator import core_validator
 from validation.proxy import sql
-from fastapi.responses import JSONResponse
-from starlette import status
-from fastapi import Response
+from validation.db.db_management import db_manager
 
 app = FastAPI(debug=False)
-
 
 app.include_router(image_router)
 app.include_router(text_router)
 
 
 async def main():
+    await db_manager.initialize()
     core_validator.start_continuous_tasks()
 
     port = validator_config.api_server_port
@@ -51,7 +51,7 @@ ENDPOINT_TO_CREDITS_USED = {
 
 
 @app.middleware("http")
-async def api_key_validator(request, call_next):
+async def api_key_validator(request: Request, call_next):
     if request.url.path in ["/docs", "/openapi.json", "/favicon.ico", "/redoc"]:
         return await call_next(request)
 
@@ -62,35 +62,32 @@ async def api_key_validator(request, call_next):
             content={"detail": "API key is missing"},
         )
 
-    with sql.get_db_connection() as conn:
-        api_key_info = sql.get_api_key_info(conn, api_key)
+    async with aiosqlite.connect(sql.DATABASE_PATH) as conn:
+        api_key_info = await sql.get_api_key_info(conn, api_key)
+        if api_key_info is None:
+            return JSONResponse(status_code=status.HTTP_401_UNAUTHORIZED, content={"detail": "Invalid API key"})
+        endpoint = request.url.path.split("/")[-1]
+        credits_required = ENDPOINT_TO_CREDITS_USED.get(endpoint, 1)
 
-    if api_key_info is None:
-        return JSONResponse(status_code=status.HTTP_401_UNAUTHORIZED, content={"detail": "Invalid API key"})
-    endpoint = request.url.path.split("/")[-1]
-    credits_required = ENDPOINT_TO_CREDITS_USED.get(endpoint, 1)
+        if api_key_info[sql.BALANCE] is not None and api_key_info[sql.BALANCE] <= credits_required:
+            return JSONResponse(
+                status_code=status.HTTP_429_TOO_MANY_REQUESTS, content={"detail": "Insufficient credits - sorry!"}
+            )
 
-    # Now check credits
-    if api_key_info[sql.BALANCE] is not None and api_key_info[sql.BALANCE] <= credits_required:
-        return JSONResponse(
-            status_code=status.HTTP_429_TOO_MANY_REQUESTS, content={"detail": "Insufficient credits - sorry!"}
-        )
-
-    # Now check rate limiting
-    with sql.get_db_connection() as conn:
-        rate_limit_exceeded = sql.rate_limit_exceeded(conn, api_key_info)
+        rate_limit_exceeded = await sql.rate_limit_exceeded(conn, api_key_info)
         if rate_limit_exceeded:
             return JSONResponse(
                 status_code=status.HTTP_429_TOO_MANY_REQUESTS, content={"detail": "Rate limit exceeded - sorry!"}
             )
 
-    response: Response = await call_next(request)
+    response = await call_next(request)
 
     if response.status_code == 200:
-        with sql.get_db_connection() as conn:
-            sql.update_requests_and_credits(conn, api_key_info, credits_required)
-            sql.log_request(conn, api_key_info, request.url.path, credits_required)
-            conn.commit()
+        async with aiosqlite.connect(sql.DATABASE_PATH) as conn:
+            await sql.update_requests_and_credits(conn, api_key_info, credits_required)
+            await sql.log_request(conn, api_key_info, request.url.path, credits_required)
+            await conn.commit()
+
     return response