+++ changelog.md
@@ -1,3 +1,12 @@
+## 4.0.6
+- Task basked checking to the orchestrator server
+- Async DB interactions
+
+## 4.0.5
+- Limiting length of synthetic generation prompts
+- Posting more to tauvision again
+- Fix for miners who sometimes get partial llm lines streamed
+- Pinning Numpy version to fix version issues
 ## 4.0.4
 - Hotfix to Capacities diminishing every period
 ## 4.0.3